+++ CHANGELOG.md
@@ -8,3 +8,4 @@
 - Get content endpoint @robgietema
 - Create content endpoint @robgietema
 - Update content endpoint @robgietema
+- Delete content endpoint @robgietema